refactor: 방이 변경된 경우 reset 리팩터링
